+++ common/src/python/identifiers/identifiers_repository.py
@@ -7,13 +7,17 @@ https://github.com/cosmicpython/code/tree/chapter_02_repository_exercise
 import abc
 import logging
 from abc import abstractmethod
+from datetime import date
 from typing import List, Optional, overload
 
 from identifiers.model import (
     CenterIdentifiers,
     GUIDField,
+    IdentifierDurationResponse,
     IdentifierList,
     IdentifierObject,
+    NACCIDField,
+    OptionalNACCADCField,
 )
 
 log = logging.getLogger(__name__)
@@ -21,6 +25,37 @@ log = logging.getLogger(__name__)
 
 class IdentifierQueryObject(CenterIdentifiers, GUIDField):
     """Query model creating objects."""
+
+
+class IdentifierUpdateObject(
+    CenterIdentifiers, GUIDField, OptionalNACCADCField, NACCIDField
+):
+    """Request model for identifier updates.
+
+    Has NACCID as string. NACCADC is optional
+    """
+
+    active: bool
+
+    @classmethod
+    def create_from_identifier(
+        cls, identifier: IdentifierObject, active: bool
+    ) -> "IdentifierUpdateObject":
+        return IdentifierUpdateObject(
+            naccid=identifier.naccid,
+            adcid=identifier.adcid,
+            ptid=identifier.ptid,
+            guid=identifier.guid,
+            naccadc=identifier.naccadc,
+            active=active,
+        )
+
+
+class DateQueryObject(NACCIDField, CenterIdentifiers):
+    """Request model for checking whether a visitdate is within the valid
+    duration for the specified center."""
+
+    visitdate: date
 
 
 class IdentifierRepository(abc.ABC):
@@ -109,6 +144,31 @@ class IdentifierRepository(abc.ABC):
 
         Returns:
           List of all identifiers in the repository or ones matching with filters
+        """
+
+    @abstractmethod
+    def add_or_update(self, identifier: IdentifierUpdateObject) -> bool:
+        """Adds/updates the Identifier record in the repository.
+
+        Args:
+          identifier: Identifier record to add/update
+
+        Returns:
+          True if add/update successful, else False
+        """
+
+    @abstractmethod
+    def check_duration(
+        self, date_query: DateQueryObject
+    ) -> Optional[IdentifierDurationResponse]:
+        """Checks whether there is a valid identifier duration record in the
+        repository matching with the visit date in query object.
+
+        Args:
+          date_query: visitdate query to validate
+
+        Returns:
+          IdentifierDurationResponse (optional) if match found, else None
         """