+++ gear/center_management/src/python/center_app/main.py
@@ -53,6 +53,10 @@ def run(
     center_roles = get_project_roles(proxy, role_names)
 
     for center in center_list:
+        if center.is_pipeline():
+            admin_group.update_center_map(adcid=center.adcid, info=center)
+            continue
+
         if new_only and center.tags and "new-center" not in center.tags:
             log.info(
                 f"new_only set to True and {center.name} does not "